+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## master
 
+- [#187](https://github.com/castle/castle-ruby/pull/187) dropped X-Client-Id from calculation of ip, drop appending default ip headers to the ip_header list config when config is provided (in that case default headers have to explicitly provided)
+
+
 ## 4.1.0 (2020-03-27)
 
 - [#184](https://github.com/castle/castle-ruby/pull/184) added Castle::API::Session which exposes Net:Http instance for reuse